+++ src/Dictum/Context.php
@@ -663,7 +663,7 @@ class Context
     public function isBlank($text): bool
     {
         if (null === ($text = $this->text($text))) {
-            return false;
+            return true;
         }
 
         return $text->isBlank();
@@ -695,10 +695,6 @@ class Context
             return 0;
         }
 
-        return $text->trim()
-            ->append(' ')
-            ->regexReplace('[^\w\s]+', '')
-            ->regexReplace('[\s]+', ' ')
-            ->countInstances(' ');
+        return $text->countWords();
     }
 }