+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.2.2 (2021-04-14)
+* Added callback support to regexReplace
+* Added Time and Number plugin interfaces
+
 ## v0.2.1 (2021-04-14)
 * Simplified label formatter
 * Added comparison helper